+++ CHANGELOG.md
@@ -1,7 +1,17 @@
-Merge home_page branch
+Implemented My Profile Page
 
 # Commit Changelog
 - Added:
+    - User's info
+        - Name
+        - XP (experience)
+        - Logo
+        - Email Address
+    - User's Insights / Stats
+        - Pending Tasks
+        - Overdue Tasks
+
+        - Current Streak
+        - Longest Streak
 - Fixed:
 - Updated:
-    - Guts destroyed while resolving conflicts